Get rid of mostly redundant Running marker-component
Process used to exist only while the underlying OS process was being executed.
Previously the Process became persistent (it would outlive its OS counterpart)
which brought a requirement for a new indication of an execution entity being
actively ran. Thats how Running component came to be.
Now since Process is back to being tyied to its OS counterpart (in terms of
their lifecycles) there is no need in Running component anymore.

+++ src/os-api.cpp
@@ -74,7 +74,9 @@ int OsApi::Exec(const std::vector<const char *> &args) {
 }
 
 void OsApi::KillProcess(Process& process) {
+  if (process.handle) {
     TinyProcessLib::Process::kill(process.handle->get_id());
+  }
 }
 
 void OsApi::ExecProcess(const std::string &shell_cmd) {
